+++ NEWS.md
@@ -3,6 +3,11 @@ CoordinateCleaner 3.0 (2023-08-31)
 ### MAJOR IMPROVEMENTS
   * Modified all functions to remove dependencies on sp, raster, and rgeos.
 
+### MINOR IMPROVEMENTS
+  * Fixed minor errors in vignettes
+  * Added an option to buffer cc_sea and cc_coun
+  * Fixed minor bugs pointed out in issues
+
 CoordinateCleaner 2.0-20 (2021-10-08)
 =========================